requirements.txt
Flask==0.10.1
Flask-SQLAlchemy==1.0
Jinja2==2.7.2
MarkupSafe==0.18
SQLAlchemy==0.9.1
Werkzeug==0.9.4
beautifulsoup4==4.3.2
gunicorn==18.0
ipython==1.1.0
itsdangerous==0.23
python-dateutil==2.2
requests==2.2.0
six==1.5.2
wsgiref==0.1.2
